pyproject.toml
[project]
name = "metadata-api"
version = "3.1"
description = ""
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Språkbanken Text", email = "sb-info@svenska.gu.se" }
]
dependencies = [
    "beautifulsoup4>=4.14.2",
    "celery[redis]>=5.5.3",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.7.0",
    "gitpython>=3.1.45",
    "jsonschema>=4.25.1",
    "lxml>=6.0.2",
    "markdown>=3.10",
    "pycountry>=24.6.1",
    "pymemcache>=4.0.0",
    "pyyaml>=6.0.3",
    "redis>=5.2.1",
    "requests>=2.32.5",
]

[tool.black]
line-length = 120

[tool.ruff]
line-length = 120
output-format = "grouped"

[tool.ruff.lint]
select = [
    "A",      # flake8-builtins
    "ANN",    # flake8-annotations
    "ARG",    # flake8-unused-arguments
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "COM",    # flake8-commas
    "D",      # pydocstyle
    "D400",   # pydocstyle: ends-in-period
    "D401",   # pydocstyle: non-imperative-mood
    "E",      # pycodestyle: errors
    "F",      # Pyflakes
    "FLY",    # flynt
    "FURB",   # refurb
    "G",      # flake8-logging-format
    "I",      # isort
    "ISC",    # flake8-implicit-str-concat
    "N",      # pep8-naming
    "PERF",   # Perflint
    "PIE",    # flake8-pie
    "PL",     # Pylint
    "PT",     # flake8-pytest-style
    "PTH",    # flake8-use-pathlib
    "Q",      # flake8-quotes
    "RET",    # flake8-return
    "RSE",    # flake8-raise
    "RUF",    # Ruff-specific rules
    "SIM",    # flake8-simplify
    "T20",    # flake8-print
    "TID",    # flake8-tidy-imports
    "UP",     # pyupgrade
    "W",      # pycodestyle: warnings
]
ignore = [
    "ANN401",  # flake8-annotations: any-type
    "B008",    # flake8-bugbear: function-call-in-default-argument
    "COM812",  # flake8-commas: missing-trailing-comma
    "E741",    # pycodestyle: ambiguous-variable-name
    "PLR09",   # Pylint: too-many-*
    "PLR1702", # Pylint: too-many-nested-blocks
    "SIM105",  # flake8-simplify: suppressible-exception
    "T201",    # flake8-print: print-statement
]

preview = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = ["uv_build>=0.9.7,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "metadata_api"
module-root = ""
